COMPLETELY split the servers. Next is to make it work with a URL e2e test
