+++ src/config/config.dev.json
@@ -3,6 +3,6 @@
   "expressServerPort": 3000,
   "mongoDbAtlasURI": "mongodb+srv://systemAdmin:hUr9bvrr4AQiDJf@rms-mongo-cluster-chess.z4pdw.mongodb.net/swiss-pairing",
   "mongoDbInMemoryURI": "",
-  "mongodbOptions": { "useCreateIndex": true, "useNewUrlParser": true, "useUnifiedTopology": true }
+  "mongodbOptions": { "useCreateIndex": true, "useNewUrlParser": true, "useUnifiedTopology": true, "useFindAndModify": false }
 }
