+++ src/index.ts
@@ -2,8 +2,18 @@
 import {IConfig} from "./config/config.interface";
 let config: IConfig = require('./config/config.dev.json');
 
+// Servers
 import { MongoAtlas } from "./server/mongodb";
-import {launchServers} from "./server/swiss-pairing";
+import app from "./server/app";
+const mongodb = new MongoAtlas(config.mongoDbAtlasURI, config.mongodbOptions);
+mongodb.connect()
+	.then(() => {
+		console.log(`MongoDB Server running`);
+		app.listen(config.expressServerPort, () => {
+			console.log(`Express HTTP Server running`);
+		});
+	})
+	.catch((err: any) => {
+		throw (err);
+	})
 
-const mongodb = new MongoAtlas(config.mongoDbAtlasURI, config.mongodbOptions)
-launchServers(mongodb);
