+++ src/model/game/game.test.ts
@@ -1,24 +1,17 @@
 // TODO: review test in detail for new architecture
-
-import express from "express";
-
 import {GAME_STATES} from "./game.constants";
+import app from "../../server/app";
 
 const request = require('supertest');
 
-import { Utils } from "../../utils/utils";
 import {TOURNAMENT_TYPE} from "../../contants/contants";
 import {AMongoDb, MongoInMemory} from "../../server/mongodb";
-import {ISwissPairingServers} from "../../server/swiss-pairings-interface";
-
-import {launchServers, stopServers} from "../../server/swiss-pairing";
 
 import {IConfig} from "../../config/config.interface";
 let config: IConfig = require('../../config/config.dev.json');
 
 
 describe('Game Entity', () => {
-	const utils = new Utils();
 	let resource = '/game';
 	let response: any;
 	let gameId: string;
@@ -30,13 +23,20 @@ describe('Game Entity', () => {
 	let gameEntity: any;
 	let gamePatch: any;
 	let mongodb: AMongoDb;
-	let swissPairingServers: ISwissPairingServers;
-	let app: express.Application;
 
 	beforeAll(async done => {
 		mongodb = new MongoInMemory(config.mongoDbInMemoryURI, config.mongodbOptions)
-		swissPairingServers = launchServers(mongodb);
-		app = swissPairingServers.applicationServer;
+		mongodb.connect()
+			.then(() => {
+				console.log(`MongoDB Server running`);
+				app.listen(config.expressServerPort, () => {
+					console.log(`Express HTTP Server running`);
+				});
+				done();
+			})
+			.catch((err: any) => {
+				done (err);
+			})
 
 		// add tournament
 		tournamentEntity = {
@@ -72,11 +72,11 @@ describe('Game Entity', () => {
 
 	afterEach(async done => {
 		await mongodb.clear();
-		done()
+		done();
 	});
 
 	afterAll(async done => {
-		stopServers(mongodb, swissPairingServers.httpServer);
+		await mongodb.close();
 		done();
 	});
 	describe('GET', () => {
