+++ src/model/round/round.test.ts
@@ -1,6 +1,6 @@
 // TODO: review test in detail for new architecture
-import express from "express";
 const request = require('supertest');
+import app from "../../server/app";
 
 import {RoundDto} from "./round.model";
 import roundDao from './round.dao';
@@ -8,8 +8,6 @@ import roundService from './round.service';
 import {ROUND_STATE} from "./round.constants";
 
 import {AMongoDb, MongoInMemory} from "../../server/mongodb";
-import {ISwissPairingServers} from "../../server/swiss-pairings-interface";
-import {launchServers, stopServers} from "../../server/swiss-pairing";
 
 import {IConfig} from "../../config/config.interface";
 let config: IConfig = require('../../config/config.dev.json');
@@ -25,13 +23,20 @@ describe('Round Entity', () => {
 	let resource = '/round';
 	let response: any;
 	let mongodb: AMongoDb;
-	let swissPairingServers: ISwissPairingServers;
-	let app: express.Application;
 
 	beforeAll(async done => {
 		mongodb = new MongoInMemory(config.mongoDbInMemoryURI, config.mongodbOptions)
-		swissPairingServers = launchServers(mongodb);
-		app = swissPairingServers.applicationServer;
+		mongodb.connect()
+			.then(() => {
+				console.log(`MongoDB Server running`);
+				app.listen(config.expressServerPort, () => {
+					console.log(`Express HTTP Server running`);
+				});
+				done();
+			})
+			.catch((err: any) => {
+				done (err);
+			})
 		done();
 	});
 
@@ -41,7 +46,7 @@ describe('Round Entity', () => {
 	});
 
 	afterAll(async done => {
-		stopServers(mongodb, swissPairingServers.httpServer);
+		await mongodb.close();
 		done();
 	});
